20 Jan 2025 - Monday
We started learning about file paths and their types. The session then progressed to CSS, where we explored the basics, starting with syntax and the box model, which includes content, border, padding, and margin. I also learned about the background-color and background-image properties, enhancing the understanding of CSS design. Additionally, I discovered the power of RGBA for setting colors with transparency, where the alpha value ranges from 0 to 1, and how RGB values range from 0 to 255. The day concluded with some topics to review and come back to tomorrow.✌